qonos: Simplify call syntax for qoCreateGraphicsPipeline()
The qoCreateGraphicsPipeline macro took two createInfo structs. All
other qoCreate* functions and macros take only one.

This patch consolidates the two struct parameters into one. This
simplifies the syntax for calling qoCreateGraphicsPipeline. And this
makes it easier to pass an inline chain of structs to the macro (which
was my real motivation).