func.depthstencil: Fix depth compare op
The yellow triangle's Z values and the depth clear value were all the
same; and the compare op was GREATER. The yellow triangle was visible in
the reference image, but that not guaranteed by GREATER because of
possible Z-fighting.

Change the op to GREATER_EQUAL, which guarantees the current reference
image.