cru/string: Remove the dangerous dummy string
string_init() and STRING_INIT initialized the string to the dummy
string:
  extern char string_dummy[1];

The dummy string was a semi-invalid state for string_t, and was
inherently unsafe.

Make the code safer by replacing the dummy string with a "short string
optimization". The main goal of the change is not optimization, though.
The goal is:
  - To ensure that statically initialized strings are always valid, and
  - To allow static initialization of non-empty strings.