func.miptree: Make template image per mipslice, not per miplevel
For color images, this is currently unnecessary because the test uses
the same template PNG image for all layers in a given level. Despite the
image reuse, we can still guarantee uniqueness among mipslice source
images. For each mipslice, the test obtains a uniquene image by
perturbing the template image's red channel as the miplevel varies and
the blue channel as the miplayer varies.

For depth images, this trick no longer works. It's not
possible to obtain a visually distinct image per mipslice by a simple
color transformation, because mipslices vary among two dimensions (level
and layer) but the depth format has only one channel. The same problem
exists for stencil images. Therfore, for depthstencil images, we will
need to use a distint template image for each miplayer.