cru/runner: Run tests in a slave process
This is a precursor to adding a runtime option that runs
one-test-per-process.

The runner now consists of two processes: master and slave. The master
forks the slave. The tests execute in the slave process. The master
collects the test results and prints their summary. The separation
ensures that test results and summary are printed even when a test
crashes its process.

For example, if you place an unavoidable abort() in vkCreateImage(),
then the test output now looks like below. Before, the output would be
truncated at the abort().

           chadv@arabesque $ crucible run
           crucible: info   : running 82 tests
           crucible: info   : ================================
           crucible: start  : func.copy.copy-buffer.large
           anv_pipeline.c:122: FINISHME: stub anv_CreatePipelineCache
  abort -> anv_device.c:2182: FINISHME: stub anv_CreateCommandPool
           crucible: info   : ================================
           crucible: info   : ran 0 tests
           crucible: info   : pass 0
           crucible: info   : fail 0
           crucible: info   : skip 0
           crucible: info   : missing 82