make: Add Makefiles to subdirs
Follow the systemd's example. Put a Makefile in each subdirectory so you
don't have to jump to the top of tree to build or clean. Jumping to
top of tree is annoying in the shell, and even more annoying in vim.
For now, like systemd, the subdir Makefiles just proxy to the main
Makefile at top of tree.