runner: Support concurrent tests when using process isolation
Add the new option -j|--jobs to crucible-run that mimics the same GNU
Make option.

The default number of jobs is the number of online CPUs.

Concurrent tests are not yet supported when using thread isolation mode.

Using -j8 reduces a full testrun from 8s to 3.5s on my machine.

This required a rewrite of the runner code.