cru: Use SPIR-V by default
The more Kristian and I have been talking about compiler stuff, the more we
want to start moving away from the GLSL compiler.  We might as well start
using SPIR-V by default now.  If this turns out to be a bad idea, it's a
single-line change to flip the switch back.