misc/glsl_scraper: Don't handle exceptions gracefully
It used to be that, of glslang couldn't compile a shader, we would
gracefully fall back to glsl.  However, since we no longer support GLSL in
our driver, this isn't really an option anymore.  Therefore, it makes no
sense to handle these errors gracefully and we should instead let the test
author know that we can't compile it.