func.miptree: Store mipslices in a cru_vec
In struct miptree, replace
    const mipslice_t *slices
    uint32_t num_slices
with
    mipslice_vec_t mipslices

This will lead to simpler code when tests for 3D textures arrive.  Using
a vector means we won't have to calculate the miptree layout twice, first when
allocating the miptree to determine the number of mipslices and again when
populating the mipslices. Instead, we can just append each mipslice to the
vector after populating it.