crucible: Convert from glslc to glslang
This essentially reverts b5a52577781ff191766f49866ce00af749cb86b7, but
it seems that glslangValidator has since become a little more
friendly. Additionally, I used python tempfile module to simplify
creating tempfiles with the required extensions.

One reason to revert to using glslangValidator is just to remove the
additional shaderc/glslc dependency. An advantage of using
glslangValidator is that the SPIR-V binary and assembly code can be
generated with a single invocation of glslangValidator.

Acked-by: Jason Ekstrand <jason@jlekstrand.net>
Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>