util: add a simple_pipeline helper function
This helper creates and submits a pipeline with a very simple vertex
shader and two triangles that cover the viewport. The user gives a
fragment shader and optionally some push constants. The idea is to avoid
boilerplate for tests that specifically test shader-only functionality.
All you need to do is create a fragment shader that renders green on
success and call this function in your test.