util/log: remove assert for pid range
This breaks crucible when running on some systems where /proc/sys/kernel/pid_max
is > UINT16_MAX (this seems to be the case on at least one system out of the box
with Debian testing). Occasionally a pid is assigned outside the range in the
assert, causing the test suite to fail.