func.multiview: Fix reference image filename
t_name now includes the queue family index number as a suffix. (For
example, ".q0" will be appended.)

It looks like func.multiview is the only test to use t_name for the
reference filename, therefore the easiest change is to just hardcode
the reference filename similar to other tests.

Fixes: e5dc897 "runner: Send queue_family_index to/from slave runner"
Reported-by: Mark Janes <mark.a.janes@intel.com>
Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>
Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>