func.compute: Avoid automatically unmapping memory
The qoMapMemory helper will register the memory to be unmapped at the
end of the test, but we also manually unmap -- because we run this in
a loop between checking and executing (when it needs to be unmapped).

This caused a second invalid vkUnmapMemory to be called.  So just use
the vkMapMemory directly as we don't need to automatic unmap.

Reviewed-by: Jordan Justen <jordan.l.justen@intel.com>