glsl_scraper: Allow for __LINE__ to be the first line of the macro
When I originally wrote the scraper, I did some testing with GCC and
found that it consistently used the last line occupied by a macro for
the value of __LINE__ within the evaluation of that macro.  However,
with GCC 9, it now uses the first line.  This commit updates the
glsl_scraper macro to allow for both behaviors.  It's rather aweful that
we rely on such details of the C preprocessor but it makes for really
nice syntax inside tests so.... Oh, well?

Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>