qonos: Don't try to provide a default blend attachment
This ends up creating a temporary short-lived stack variable which
cannot persist past it's immediate surroundings.  This works ok if
you're creating it just to pass into a function but in this case, we're
creating it, stashing it in a struct, doing more stuff, and then passing
it into the function.  The result is that it doesn't have a well-defined
life time and, even if it did, it would probably still be restricted to
the if statement in which we set up the blend state.  On newer GCC 9,
this causes qoCreateGraphicsPipeline to explode.  Instead, create the
attachment on the stack just like everything else.

Reviewed-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>