Show certain VK_EXT_debug_report messages only when --verbose
When VK_EXT_debug_report is available, `crucible run` prints a lot of
information and debug messages.  Change it to by default print only
the warnings and errors, and then print everything when `--verbose`
options is used.

The info messages include listings of all available extensions which
makes the output of very noisy.