tests: remove func.sync.semaphore-fd.opaque-fd
This tests is not valid because it relies on implementation specific
details.

When queueing an operations into a VkQueue, the semaphores operations
associated with a given VkSubmitInfo are going to be executed at a
given point in that over which the application has no control (this
depends on how the implementation processes operations in the queue).
That means that for a semaphore to be signaled/waited/signaled/etc...
multiple time we need external synchronization to ensure that a wait
applies to a particular signal operations.

This test was queueing many wait/signal operations on the same
semaphore without any synchronization between them, leading to
undefined behavior.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Reviewed-by: Bas Nieuwenhuizen <bas@basnieuwenhuizen.nl>