framework: Only create one command pool per queue family
We only need to create one command pool per queue family, not per
queue. Each command buffer created on a command pool for a queue
family can be submitted to any queue on that queue family.

Here we fill in an array (t->vk.cmd_pool) where every queue can access
an appropriate command pool for it's queue family. The queues that
share the same family get assigned the same command pool in the array.

Recommended-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Ref: https://gitlab.freedesktop.org/mesa/crucible/-/merge_requests/77#note_782934
Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>
Reviewed-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>