qonos: Do not rely on VK_DYNAMIC_STATE_RANGE_SIZE
This was removed in newer Vulkan headers, because actual values added
by extensions are way beyond the small integers, e.g.
VK_DYNAMIC_STATE_FRAGMENT_SHADING_RATE_KHR value is 1000226000.

This was used to setup pDynamicStates in case none was provided, in
practice just avoiding a bit of boilerplate.  Because the bit flag in
QoExtraGraphicsPipelineCreateInfo limited to 32 bits, we can use that
limit.

Reviewed-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>